+++ packages/cli/src/lib/codegen/index.ts
@@ -1,3 +1,4 @@
 export * from "./CodeGenerator";
 export * from "./CodegenOverrides";
 export * from "./ScriptCodeGenerator";
+export * from "./WasmEmbed";