+++ packages/js/manifests/polywrap/src/formats/polywrap.plugin/migrators/0.4.0_to_0.5.0.ts
@@ -1,0 +1,9 @@
+import { PluginManifest as OldManifest } from "../0.4.0";
+import { PluginManifest as NewManifest } from "../0.5.0";
+
+export function migrate(migrate: OldManifest): NewManifest {
+  return {
+    ...migrate,
+    format: "0.5.0",
+  };
+}